requirements.txt
pydantic
openai
crewai
crewai-tools
landingai-ade
python-dotenv
langchain-openai
langchain-groq
